# src/controllers/v3/openapi.yaml
components:
...
    schemas:
...
        BatchResult:
            properties:
                status:
                    type: number
                    format: double
                result:
                    type: string
                reason:
                    type: string
                id:
                    type: string
                ownedBy:
                    type: string
                childResults:
                    items:
                        $ref: '#/components/schemas/BatchResult'
                    type: array
            required:
                - status
                - result
            type: object
            additionalProperties: false
...
        Gateway:
            properties:
                gatewayId:
                    type: string
                displayName:
                    type: string
            type: object
            additionalProperties: false
...
        GatewayServiceRefID:
            type: string
...
        DraftDatasetRefID:
            type: string
        LegalRefID:
            type: string
        CredentialIssuerRefID:
            type: string
        Environment:
            properties:
                appId:
                    type: string
                name:
                    type: string
                    enum:
                        - dev
                        - test
                        - prod
                        - sandbox
                        - other
                active:
                    type: boolean
                approval:
                    type: boolean
                flow:
                    type: string
                    enum:
                        - public
                        - protected-externally
                        - authorization-code
                        - client-credentials
                        - kong-acl-only
                        - kong-api-key-only
                        - kong-api-key-acl
                additionalDetailsToRequest:
                    type: string
                services:
                    items:
                        $ref: '#/components/schemas/GatewayServiceRefID'
                    type: array
                legal:
                    $ref: '#/components/schemas/LegalRefID'
                credentialIssuer:
                    $ref: '#/components/schemas/CredentialIssuerRefID'
            type: object
            additionalProperties: false
            example:
                name: dev
                active: false
                approval: false
                flow: public
                appId: '00000000'
        Product:
            properties:
                appId:
                    type: string
                name:
                    type: string
                description:
                    type: string
                gatewayId:
                    type: string
                dataset:
                    $ref: '#/components/schemas/DraftDatasetRefID'
                environments:
                    items:
                        $ref: '#/components/schemas/Environment'
                    type: array
            type: object
            additionalProperties: false
            example:
                name: my-new-product
                appId: '000000000000'
                environments:
                    -
                        name: dev
                        active: false
                        approval: false
                        flow: public
                        appId: '00000000'
    securitySchemes:
        jwt:
            type: oauth2
            description: 'Authz Client Credential'
            flows:
                clientCredentials:
                    tokenUrl: 'https://token_endpoint'
                    scopes: {}
        portal:
            type: http
            description: 'Authz Portal Login'
            scheme: bearer
            bearerFormat: JWT
        openid:
            type: openIdConnect
            description: 'OIDC Login'
            openIdConnectUrl: 'https://well_known_endpoint'
info:
    title: 'APS Directory API'
    version: 3.0.0
    description: 'API Services Portal by BC Gov API Programme Services'
    license:
        name: MIT
    contact:
        name: 'BC Gov APS'
openapi: 3.0.0
paths:
...
    '/gateways/{gatewayId}':
        get:
            operationId: namespace-profile
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Gateway'
            description: "Get details about the gateway, such as permissions for what the gateway is setup with.\n> `Required Scope:` Namespace.Manage"
            summary: 'Gateway Summary'
            tags:
                - Gateways
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
        delete:
            operationId: delete-namespace
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Gateway'
            description: "Delete the gateway\n> `Required Scope:` Namespace.Manage"
            summary: 'Delete Gateway'
            tags:
                - Gateways
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: force
                    required: false
                    schema:
                        default: false
                        type: boolean
...
    '/gateways/{gatewayId}/products':
        put:
            operationId: put-product
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchResult'
            description: "Manage Products for APIs that will appear on the API Directory\n> `Required Scope:` Namespace.Manage"
            summary: 'Manage Products'
            tags:
                - Products
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Product'
        get:
            operationId: get-products
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Product'
                                type: array
            description: "Get Products describing APIs that will appear on the API Directory\n> `Required Scope:` Namespace.Manage"
            summary: 'Get Products'
            tags:
                - Products
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters: []
    '/gateways/{gatewayId}/products/{appId}':
        delete:
            operationId: delete-product
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchResult'
            description: "Delete a Product\n> `Required Scope:` Namespace.Manage"
            summary: 'Manage Products'
            tags:
                - Products
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: appId
                    required: true
                    schema:
                        type: string
    '/gateways/{gatewayId}/environments/{appId}':
        delete:
            operationId: delete-product-environment
            responses:
                '204':
                    description: 'No content'
            description: "Delete a Product Environment\n> `Required Scope:` Namespace.Manage"
            summary: 'Delete a Product Environment'
            tags:
                - Products
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: appId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: force
                    required: false
                    schema:
                        default: false
                        type: boolean
servers:
    -
        url: /ds/api/v3
...